Adjusted the positions of the two passes, placing them after the LoweTileOp Pass; set vector_load_bits_max_ to 1024 when the target is Summio; and made some other minor adjustments.
