+++ src/transform/loop_vectorize.cc
@@ -175,6 +175,8 @@ public:
         VectorizeFindMemoryAccess::MaySupportVectorize256(node)) {
       vector_load_bits_max_ = initial_vector_size_ = loop_extent_vector_size_ =
           256;
+    } else if (tvm::tl::TargetIsSunmmio(Target::Current(false))) {
+      vector_load_bits_max_ = vector_size_ = 1024;
     } else {
       vector_load_bits_max_ = initial_vector_size_ = loop_extent_vector_size_ =
           128;
