+++ tilelang/engine/phase.py
@@ -157,14 +157,9 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
         IRModule: The transformed module, ready for target-specific optimization passes.
     """
     mod = tir.transform.BindTarget(target)(mod)
-
     if should_force_let_inline():
         # Force-let inline whenever the pass config requests it.
         mod = tilelang.transform.LetInline()(mod)
-    # read TileView metadata and attach them to Tiles loops.
-    mod = tilelang.transform.LegalizeTilesLoop()(mod)
-    # add two for loops for tile loops
-    mod = tilelang.transform.TilesLoop()(mod)
     # Add wrapper for single buf store
     mod = tilelang.transform.AddWrapperForSingleBufStore()(mod)
     # Normalize negative indices to canonical non-negative form
@@ -186,6 +181,10 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
     LayoutVisual(mod)
     # Lower high-level tile operations to low-level operations
     mod = tilelang.transform.LowerTileOp()(mod)
+    # read TileView metadata and attach them to Tiles loops.
+    mod = tilelang.transform.LegalizeTilesLoop()(mod)
+    # trans the tiles loop into two layers of loops, the inner one is a vectorized loop and the outer one is a Serial loop
+    mod = tilelang.transform.TilesLoop()(mod)
     # Lower l2 persistent map
     mod = tilelang.transform.LowerL2Persistent()(mod)
     # Decouple type cast vectorization constraints before vectorization
